+++ src/jsx/components/library/song/List.jsx
@@ -5,13 +5,13 @@ import SongEntry from './Entry'
 import ListingFetchWrapper from 'components/generics/ListingFetchWrapper'
 import Navigator from 'components/generics/Navigator'
 import { songStatePropType } from 'reducers/library'
-import { playlistStatePropType } from 'reducers/playlist'
+import { playlistEntriesStatePropType } from 'reducers/playlist'
 import { playlistAppDigestPropType } from 'reducers/playlistApp'
 
 class SongList extends Component {
     static propTypes = {
         songState: songStatePropType.isRequired,
-        playlistState: playlistStatePropType.isRequired,
+        playlistEntriesState: playlistEntriesStatePropType.isRequired,
         playlistAppDigest: playlistAppDigestPropType.isRequired,
         location: PropTypes.object.isRequired,
     }
@@ -25,24 +25,16 @@ class SongList extends Component {
          * and currently playing song
          */
 
-        const currentTime = new Date().getTime()
         let remainingTime = 0
         const playerStatus = this.props.playlistAppDigest.data.player_status
-        const playlistInfo = {}
 
         // First, if a song is playing,
         // set remainingTime to the duration until end of the song
 
         if (playerStatus.playlist_entry) {
             // Player is playing a song
-            // get playing id
-            // and add remaining duration
+            // add remaining duration
             const playingId = playerStatus.playlist_entry.song.id
-            playlistInfo[playingId] = {
-                isPlaying: true,
-                owner: playerStatus.playlist_entry.owner
-            }
-
             remainingTime = playerStatus.playlist_entry.song.duration - playerStatus.timing
         }
 
@@ -51,9 +43,12 @@ class SongList extends Component {
         // generate a map with song id as key
         // and time + owner as value
 
-        for (let entry of this.props.playlistState.data.playlistEntries) {
-            if (!playlistInfo[entry.song.id]) {
-                playlistInfo[entry.song.id] = {
+        const currentTime = new Date().getTime()
+        const queueInfo = {}
+
+        for (let entry of this.props.playlistEntriesState.data.playlistEntries) {
+            if (!queueInfo[entry.song.id]) {
+                queueInfo[entry.song.id] = {
                     timeOfPlay: currentTime + remainingTime * 1000,
                     owner: entry.owner
                 }
@@ -70,7 +65,7 @@ class SongList extends Component {
                 <SongEntry
                         key={song.id}
                         song={song}
-                        playlistInfo={playlistInfo[song.id]}
+                        queueInfo={queueInfo[song.id]}
                         location={this.props.location}
                     />
             ))
@@ -100,7 +95,7 @@ class SongList extends Component {
 
 const mapStateToProps = (state) => ({
     songState: state.library.song,
-    playlistState: state.playlistApp.playlist,
+    playlistEntriesState: state.playlistApp.entries,
     playlistAppDigest: state.playlistApp.digest,
 })