+++ src/jsx/components/playlistApp/player/Player.jsx
@@ -91,11 +91,6 @@ class Player extends Component {
                                 icon={player_manage.pause ? 'play' : 'pause'}
                                 iconDisabled="stop"
                             />
-                        </IsPlaylistManagerOrOwner>
-                        <IsPlaylistManagerOrOwner
-                            object={player_status.playlist_entry}
-                            disable
-                        >
                             <ManageButton
                                 manageStatus={sendPlayerCommandsStatusSafe.skip}
                                 onClick={() =>